+++ config/drawerEntries.ts
@@ -5,6 +5,7 @@ import {AclPermissions} from "~/functions/acl/enums/acl.permissions";
 import {ClubPermissions} from "~/functions/acl/enums/club.permissions";
 import {CalculatorPermissions} from "~/functions/acl/enums/calculator.permissions";
 import permissions from "~/functions/permissions";
+import {SettingPermissions} from "~/functions/acl/enums/setting.permissions";
 
 export interface DrawerEntry {
   id?: string
@@ -158,4 +159,14 @@ export default function (context: Vue): DrawerEntry[] {
     }
   ]
 
+}
+
+export function BottomDrawerEntries(context: Vue): DrawerEntry[] {
+  return [{
+    id: 'settings',
+    text: 'settings',
+    icon: 'mdi-cog',
+    link: '/settings',
+    permissions: [SettingPermissions.SETTINGS_ALL_READ, SettingPermissions.SETTINGS_SELF_READ]
+  }]
 }