+++ functions/acl/enums/setting.permissions.ts
@@ -1,0 +1,6 @@
+export enum SettingPermissions {
+  SETTINGS_ALL_READ = "settings.all:read",
+  SETTINGS_ALL_WRITE = "settings.all:write",
+  SETTINGS_SELF_READ = "settings.self:read",
+  SETTINGS_SELF_WRITE = "settings.self:write",
+}