+++ plugins/vuex-persist.js
@@ -4,6 +4,7 @@ export default ({ store }) => {
   new VuexPersistence({
     reducer: (state) => ({
       auth: state.auth,
+      settings: state.settings,
       enums: state.enums,
       i18n: state.i18n,
     }),