+++ static/changelog.js
@@ -1,17 +1,31 @@
-module.exports = `<h2 id="v189small28042020small">v 1.8.9 - <small>28/04/2020</small></h2>
+module.exports = `<h2 id="v1810small21052020small">v 1.8.10 - <small>21/05/2020</small></h2>
 <h3 id="miglioramenti">Miglioramenti</h3>
+<ul>
+<li>Sistemato bug che non permetteva il download delle integrazioni.</li>
+<li>Aggiornati componenti interni dell'applicazione (campo valuta, tabelle, pagina richieste)</li>
+</ul>
+<h3 id="aggiunte">Aggiunte</h3>
+<ul>
+<li>Possibilità agli admin di creare richieste per conto del cliente.</li>
+<li>Colonna gold in tabella utenti che indica il Club Pack di un utente.</li>
+<li>Sezione "Impostazioni" per gestire le configurazioni di tutta l'applicazione.</li>
+<li>Impostazioni per settare la fascia oraria in cui è possibile operara all'interno del sistema.</li>
+<li>Impostazione che permette di forzare tutta l'app (indipendentemente dalla fascia oraria e dai diritti dell'utente) in modalità "sola lettura".</li>
+</ul>
+<h2 id="v189small28042020small">v 1.8.9 - <small>28/04/2020</small></h2>
+<h3 id="miglioramenti-1">Miglioramenti</h3>
 <ul>
 <li>Aggiornate icone che indicano le note all'interno delle tabelle</li>
 <li>Sistemato problema con il calcolo dei brite disponibili nel profilo del Club</li>
 <li>Permette di inserire caratteri alfanumerici come numero tessera socio club </li>
 </ul>
-<h3 id="aggiunte">Aggiunte</h3>
+<h3 id="aggiunte-1">Aggiunte</h3>
 <ul>
 <li>Aggiunta possibilità per gli agenti di creare richiesta automatiche o ricorsive di riscossione delle proprie provvigioni.</li>
 <li>Permette ai superadmin di modificare le email degli utenti</li>
 </ul>
 <h2 id="v188small15042020small">v 1.8.8 - <small>15/04/2020</small></h2>
-<h3 id="miglioramenti-1">Miglioramenti</h3>
+<h3 id="miglioramenti-2">Miglioramenti</h3>
 <ul>
 <li>Anagrafica, campi obbligatori solo se non bozza.</li>
 <li>Aggiornata tabella richieste da evadere in dashboard</li>
@@ -20,51 +34,51 @@ module.exports = `<h2 id="v189small28042020small">v 1.8.9 - <small>28/04/2020</s
 <li>Aggiunto simbolo diverso per indicare i campi che momentaneamente non sono obbligatori perchè utente in bozza, ma che
 lo saranno quando si cercherà di attivare l'utente</li>
 </ul>
-<h3 id="aggiunte-1">Aggiunte</h3>
+<h3 id="aggiunte-2">Aggiunte</h3>
 <ul>
 <li>Aggiunto possibilità di trasferimento provvigioni da agente di riferimento a subagente</li>
 <li>In fase di approvazione richiesta, permette di cambiare l'importo di quest'ultima</li>
 <li>Aggiunti blocchi riepilogativi in elenco utenti relativi agli stati</li>
 </ul>
 <h2 id="v187small12042020small">v 1.8.7 - <small>12/04/2020</small></h2>
-<h3 id="miglioramenti-2">Miglioramenti</h3>
+<h3 id="miglioramenti-3">Miglioramenti</h3>
 <ul>
 <li>Aggiornata pagina del profilo degli utenti</li>
 <li>Abbassato l'età minima per apertura account a 14 anni</li>
 <li>Mostra i log delle versioni solo per gli utenti amministrativi</li>
 </ul>
-<h3 id="aggiunte-2">Aggiunte</h3>
+<h3 id="aggiunte-3">Aggiunte</h3>
 <ul>
 <li>Aggiunto pulsante per passare dall'anagrafica al profilo e viceversa</li>
 <li>Popup termini e condizioni nella pagina del contratto</li>
 </ul>
 <h2 id="v180small01042020small">v 1.8.0 - <small>01/04/2020</small></h2>
-<h3 id="miglioramenti-3">Miglioramenti</h3>
+<h3 id="miglioramenti-4">Miglioramenti</h3>
 <ul>
 <li>Aggiornata logica di creazione ed attivazione account.</li>
 <li>Implementata logica log firma contratto.</li>
 <li>Rimosso riferimenti a date nei preventivi</li>
 </ul>
-<h3 id="aggiunte-3">Aggiunte</h3>
+<h3 id="aggiunte-4">Aggiunte</h3>
 <ul>
 <li>Pulsante rimozione manuale Brite</li>
 <li>Pulsante aggiunta manuale Provvigioni</li>
 <li>Pulsante per reinviare il contratto da firmare</li>
 </ul>
 <h2 id="v173small13032020small">v 1.7.3 - <small>13/03/2020</small></h2>
-<h3 id="miglioramenti-4">Miglioramenti</h3>
+<h3 id="miglioramenti-5">Miglioramenti</h3>
 <ul>
 <li>Risolti problemi richieste di riscossione in mobile</li>
 <li>Sistemato menu inferiore su iPhone lasciando spazio per la home line</li>
 <li>Menu "altro" in mobile, non mostra più gli elementi già visibili nella barra inferiore</li>
 <li>Calcolatrice spostato aggiunta movimenti direttamente nella tabella</li>
 </ul>
-<h3 id="aggiunte-4">Aggiunte</h3>
+<h3 id="aggiunte-5">Aggiunte</h3>
 <ul>
 <li>Tabella con elenco contratti in attesa di firma nella dashboard admin</li>
 </ul>
 <h2 id="v170small23022020small">v 1.7.0 - <small>23/02/2020</small></h2>
-<h3 id="aggiunte-5">Aggiunte</h3>
+<h3 id="aggiunte-6">Aggiunte</h3>
 <ul>
 <li>Campo "IBAN", "BIC" e "Note" nel report excel delle richieste</li>
 <li>Aggiunto campo "Tipologia Provvigioni squadra" in caso di agente con subagenti</li>
@@ -75,7 +89,7 @@ lo saranno quando si cercherà di attivare l'utente</li>
 <li>Aggiunta sezione dedicata ai ruoli e permessi di tutti gli utenti</li>
 <li>Comunicazioni - Aggiunta tab dedicata al Global Club</li>
 </ul>
-<h3 id="miglioramenti-5">Miglioramenti</h3>
+<h3 id="miglioramenti-6">Miglioramenti</h3>
 <ul>
 <li>Spostato impostazioni agente (anagrafica utente) in tab dedicata</li>
 <li>Aggiunta colonna Brite Parziali nella calcolatrice</li>
@@ -83,47 +97,47 @@ lo saranno quando si cercherà di attivare l'utente</li>
 <li>Modale che mostra i clienti permette di vedere anche i sottoclienti di un subagente</li>
 </ul>
 <h2 id="v163small06022020small">v 1.6.3 - <small>06/02/2020</small></h2>
-<h3 id="aggiunte-6">Aggiunte</h3>
+<h3 id="aggiunte-7">Aggiunte</h3>
 <ul>
 <li>Pulsanti per scaricare il preventivo corrente</li>
 <li>Menù per il download delle ricevute nell'elenco movimenti</li>
 </ul>
 <h2 id="v162small28012020small">v 1.6.2 - <small>28/01/2020</small></h2>
-<h3 id="aggiunte-7">Aggiunte</h3>
+<h3 id="aggiunte-8">Aggiunte</h3>
 <ul>
 <li>Barra di ricerca nell'elenco degli utenti</li>
 <li>Barra di navigazione in mobile</li>
 <li>Calcolatrice per preventivi</li>
 </ul>
-<h3 id="miglioramenti-6">Miglioramenti</h3>
+<h3 id="miglioramenti-7">Miglioramenti</h3>
 <ul>
 <li>Lettura tabelle in mobile </li>
 </ul>
 <h2 id="v161small05012021small">v 1.6.1 - <small>05/01/2021</small></h2>
-<h3 id="aggiunte-8">Aggiunte</h3>
+<h3 id="aggiunte-9">Aggiunte</h3>
 <ul>
 <li>Transizioni al cambio delle pagine</li>
 <li>Aggiunta lista dei cambiamenti per ogni versione</li>
 <li>Email nel caso di notifica in app non letta entro 30 minuti</li>
 </ul>
-<h3 id="miglioramenti-7">Miglioramenti</h3>
+<h3 id="miglioramenti-8">Miglioramenti</h3>
 <ul>
 <li>Grafica splash screen iniziale</li>
 <li>Cambiato popup comunicazioni chat, da schermo intero a popup normale</li>
 </ul>
 <h2 id="v160small05012021small">v 1.6.0 - <small>05/01/2021</small></h2>
-<h3 id="aggiunte-9">Aggiunte</h3>
+<h3 id="aggiunte-10">Aggiunte</h3>
 <ul>
 <li>Sezione "Gold / Global club" in anagrafica utente</li>
 <li>Pulsante "Riscossione gold" in pagina richieste</li>
 <li>Modale per "Riscossione Gold" per riscossione interessi o prelievo oro fisico</li>
 </ul>
-<h3 id="miglioramenti-8">Miglioramenti</h3>
+<h3 id="miglioramenti-9">Miglioramenti</h3>
 <ul>
 <li>Grafico dashboard agenti e clienti - ora mostra anche 6 mesi futuri </li>
 </ul>
 <h2 id="v150small04012021small">v 1.5.0 - <small>04/01/2021</small></h2>
-<h3 id="aggiunte-10">Aggiunte</h3>
+<h3 id="aggiunte-11">Aggiunte</h3>
 <ul>
 <li>Campo "metodo pagamento iniziale"</li>
 <li>Campo "investimento iniziale in oro"</li>
@@ -132,7 +146,7 @@ lo saranno quando si cercherà di attivare l'utente</li>
 <li>Pulsante per scaricare la contabile di un versamento</li>
 <li>Pulsante per scaricare il report dei bonifici da effettuare</li>
 </ul>
-<h3 id="miglioramenti-9">Miglioramenti</h3>
+<h3 id="miglioramenti-10">Miglioramenti</h3>
 <ul>
 <li>Corretta validazione tipologia utente in anagrafica</li>
 <li>Aggiornata la grafica del menu laterale e barra di navigazione</li>